refactor(inputs.mem): Clean up extended memory follow-up nits
Post-merge review follow-ups for #18532:
- Flip Gather's inverted err-check and use maps.Copy for the extended
  field merge.
- Group CollectExtended and Log into a single config block in the struct.
- Split the Linux- and Windows-specific extended tests into
  mem_linux_test.go and mem_windows_test.go using build tags, matching
  the pattern used by diskio and zfs, and drop the runtime.GOOS skips.
- Share the base expected field map in the Linux extended test so the
  two table cases only specify what actually differs.
