+++ plugins/inputs/mem/mem_test.go
@@ -1,8 +1,6 @@
 package mem
 
 import (
-	"path/filepath"
-	"runtime"
 	"testing"
 	"time"
 
@@ -117,195 +115,4 @@ func TestMemStats(t *testing.T) {
 	}
 
 	testutil.RequireMetricsEqual(t, expected, acc.GetTelegrafMetrics(), testutil.IgnoreTime())
-}
-
-func TestMemStatsCollectExtended(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skip("Skipping Linux-specific extended memory test")
-	}
-
-	tests := []struct {
-		name        string
-		testdataDir string
-		expected    []telegraf.Metric
-	}{
-		{
-			name:        "normal",
-			testdataDir: "normal",
-			expected: []telegraf.Metric{
-				metric.New(
-					"mem",
-					map[string]string{},
-					map[string]interface{}{
-						"total":             uint64(12400),
-						"available":         uint64(7600),
-						"used":              uint64(5000),
-						"used_percent":      100 * float64(5000) / float64(12400),
-						"available_percent": 100 * float64(7600) / float64(12400),
-						"free":              uint64(1235),
-						"active":            uint64(0),
-						"buffered":          uint64(0),
-						"cached":            uint64(0),
-						"commit_limit":      uint64(0),
-						"committed_as":      uint64(0),
-						"dirty":             uint64(0),
-						"high_free":         uint64(0),
-						"high_total":        uint64(0),
-						"huge_pages_free":   uint64(0),
-						"huge_page_size":    uint64(0),
-						"huge_pages_total":  uint64(0),
-						"inactive":          uint64(0),
-						"low_free":          uint64(0),
-						"low_total":         uint64(0),
-						"mapped":            uint64(0),
-						"page_tables":       uint64(0),
-						"shared":            uint64(0),
-						"slab":              uint64(0),
-						"sreclaimable":      uint64(0),
-						"sunreclaim":        uint64(0),
-						"swap_cached":       uint64(0),
-						"swap_free":         uint64(0),
-						"swap_total":        uint64(0),
-						"vmalloc_chunk":     uint64(0),
-						"vmalloc_total":     uint64(0),
-						"vmalloc_used":      uint64(0),
-						"write_back_tmp":    uint64(0),
-						"write_back":        uint64(0),
-						"active_anon":       uint64(5765169152),
-						"inactive_anon":     uint64(1082245120),
-						"active_file":       uint64(3535425536),
-						"inactive_file":     uint64(4421992448),
-						"unevictable":       uint64(143360),
-						"percpu":            uint64(5767168),
-					},
-					time.Unix(0, 0),
-					telegraf.Gauge,
-				),
-			},
-		},
-		{
-			name:        "missing fields",
-			testdataDir: "missing_fields",
-			expected: []telegraf.Metric{
-				metric.New(
-					"mem",
-					map[string]string{},
-					map[string]interface{}{
-						"total":             uint64(12400),
-						"available":         uint64(7600),
-						"used":              uint64(5000),
-						"used_percent":      100 * float64(5000) / float64(12400),
-						"available_percent": 100 * float64(7600) / float64(12400),
-						"free":              uint64(1235),
-						"active":            uint64(0),
-						"buffered":          uint64(0),
-						"cached":            uint64(0),
-						"commit_limit":      uint64(0),
-						"committed_as":      uint64(0),
-						"dirty":             uint64(0),
-						"high_free":         uint64(0),
-						"high_total":        uint64(0),
-						"huge_pages_free":   uint64(0),
-						"huge_page_size":    uint64(0),
-						"huge_pages_total":  uint64(0),
-						"inactive":          uint64(0),
-						"low_free":          uint64(0),
-						"low_total":         uint64(0),
-						"mapped":            uint64(0),
-						"page_tables":       uint64(0),
-						"shared":            uint64(0),
-						"slab":              uint64(0),
-						"sreclaimable":      uint64(0),
-						"sunreclaim":        uint64(0),
-						"swap_cached":       uint64(0),
-						"swap_free":         uint64(0),
-						"swap_total":        uint64(0),
-						"vmalloc_chunk":     uint64(0),
-						"vmalloc_total":     uint64(0),
-						"vmalloc_used":      uint64(0),
-						"write_back_tmp":    uint64(0),
-						"write_back":        uint64(0),
-						"active_anon":       uint64(5765169152),
-						"inactive_anon":     uint64(1082245120),
-						"active_file":       uint64(3535425536),
-						"inactive_file":     uint64(4421992448),
-						"unevictable":       uint64(0),
-						"percpu":            uint64(0),
-					},
-					time.Unix(0, 0),
-					telegraf.Gauge,
-				),
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			hostProc, err := filepath.Abs(filepath.Join("testdata", tt.testdataDir, "proc"))
-			require.NoError(t, err)
-			t.Setenv("HOST_PROC", hostProc)
-
-			var mps psutil.MockPS
-			defer mps.AssertExpectations(t)
-
-			vms := &mem.VirtualMemoryStat{
-				Total:     12400,
-				Available: 7600,
-				Used:      5000,
-				Free:      1235,
-			}
-
-			mps.On("VMStat").Return(vms, nil)
-			plugin := &Mem{
-				ps:              &mps,
-				CollectExtended: true,
-			}
-			require.NoError(t, plugin.Init())
-
-			var acc testutil.Accumulator
-			require.NoError(t, plugin.Gather(&acc))
-
-			testutil.RequireMetricsEqual(t, tt.expected, acc.GetTelegrafMetrics(), testutil.IgnoreTime())
-		})
-	}
-}
-
-func TestMemStatsCollectExtendedWindows(t *testing.T) {
-	if runtime.GOOS != "windows" {
-		t.Skip("Skipping Windows-specific extended memory test")
-	}
-
-	plugin := &Mem{
-		ps:              psutil.NewSystemPS(),
-		CollectExtended: true,
-	}
-	require.NoError(t, plugin.Init())
-
-	var acc testutil.Accumulator
-	require.NoError(t, plugin.Gather(&acc))
-
-	expected := []telegraf.Metric{
-		metric.New(
-			"mem",
-			map[string]string{},
-			map[string]interface{}{
-				"total":             uint64(0),
-				"available":         uint64(0),
-				"used":              uint64(0),
-				"used_percent":      float64(0),
-				"available_percent": float64(0),
-				"commit_limit":      uint64(0),
-				"commit_total":      uint64(0),
-				"virtual_total":     uint64(0),
-				"virtual_avail":     uint64(0),
-				"phys_total":        uint64(0),
-				"phys_avail":        uint64(0),
-				"page_file_total":   uint64(0),
-				"page_file_avail":   uint64(0),
-			},
-			time.Unix(0, 0),
-			telegraf.Gauge,
-		),
-	}
-	testutil.RequireMetricsStructureEqual(t, expected, acc.GetTelegrafMetrics(), testutil.IgnoreTime())
 }
